Update ComabtSystem is DONE
Combat System is DONE for now
Bugs
- on Item Card, make  duplicate went drag out from battle deck

+++ Assets/Interface/IDamageable.cs
@@ -2,6 +2,6 @@ using UnityEngine;
 
 public interface IDamageable
 {
-    int healtPoints { get; set; }
-    public void TakeDamage(int damage);
+    float healtPoints { get; set; }
+    public void TakeDamage(float damage);
 }
